Incluido campo loja na relação de coletas

+++ backend/routes/collects.js
@@ -3,6 +3,7 @@ import jwt from "jsonwebtoken";
 import 'dotenv/config';
 import { companiesArray } from "./companies&products.js";
 import { productsArray } from "./companies&products.js";
+import { firmArray } from "./companies&products.js";
 
 function logTime(label) {
     console.time(label);
@@ -28,6 +29,7 @@ const getOptions = {
                 volume: { type: "integer" },
                 weight: { type: "integer" },
                 order: { type: "integer" },
+                loja: { type: "string"},
               },
               required: ["id", "company", "date", "product", "username"],
             }
@@ -88,8 +90,13 @@ async function collects(app, _) {
                         if (volumeAndWeightRegex.test(req.body.data.volume)) {
                             if (volumeAndWeightRegex.test(req.body.data.weight)) {
                                 if (orderRegex.test(req.body.data.order)) {
-                                    await db.query(`INSERT INTO coletasaprovar (company, date, product, username, volume, weight, "order") VALUES($1, $2, $3, $4, $5, $6, $7)`, [req.body.data.company, req.body.data.date, req.body.data.product, req.body.collaborator.name, req.body.data.volume, req.body.data.weight, req.body.data.order]);
-                                    res.status(201).send("ok");
+                                    if (firmArray.includes(req.body.data.loja)) {
+                                        await db.query(`INSERT INTO coletasaprovar (company, date, product, username, volume, weight, "order", loja) VALUES($1, $2, $3, $4, $5, $6, $7, $8)`, [req.body.data.company, req.body.data.date, req.body.data.product, req.body.collaborator.name, req.body.data.volume, req.body.data.weight, req.body.data.order, req.body.data.loja]);
+                                        res.status(201).send("ok");
+                                    }
+                                    if (!firmArray.includes(req.body.data.loja)) {
+                                        res.status(404).send("Coloque um nome de loja válido");
+                                    }
                                 }
                                 if (orderRegex.test(req.body.data.order)) {
                                     res.status(404).send("Pedido deve ser positivo, inteiro e no máximo 12 dígitos");
@@ -129,12 +136,12 @@ async function collects(app, _) {
         try {
             if (req.body.filter == "coletas") {
                 const result = await db.query("DELETE FROM coletas WHERE id = ($1) RETURNING *", [req.body.itemId]);
-                await db.query(`INSERT INTO coletasfeitas (company, date, product, username, volume, weight, "order") VALUES($1, $2, $3, $4, $5, $6, $7)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order]);
+                await db.query(`INSERT INTO coletasfeitas (company, date, product, username, volume, weight, "order", loja) VALUES($1, $2, $3, $4, $5, $6, $7, $8)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order, result.rows[0].loja]);
                 res.status(200).send("ok");
             }
             if (req.body.filter == "coletasaprovar") {
                 const result = await db.query("DELETE FROM coletasaprovar WHERE id = ($1) RETURNING *", [req.body.itemId]);
-                await db.query(`INSERT INTO coletas (company, date, product, username, volume, weight, "order") VALUES($1, $2, $3, $4, $5, $6, $7)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order]);
+                await db.query(`INSERT INTO coletas (company, date, product, username, volume, weight, "order", loja) VALUES($1, $2, $3, $4, $5, $6, $7, $8)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order, result.rows[0].loja]);
                 res.status(200).send("ok");
             }
         }
@@ -152,12 +159,12 @@ async function collects(app, _) {
         try {
             if (req.body.filter == "coletas") {
                 const result = await db.query("DELETE FROM coletas WHERE id = ($1) RETURNING *", [req.body.itemId]);
-                await db.query(`INSERT INTO coletasdeletadas (company, date, product, username, volume, weight, "order") VALUES($1, $2, $3, $4, $5, $6, $7)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order]);
+                await db.query(`INSERT INTO coletasdeletadas (company, date, product, username, volume, weight, "order", loja) VALUES($1, $2, $3, $4, $5, $6, $7, $8)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order, result.rows[0].loja]);
                 res.status(200).send("ok");
             }
             if (req.body.filter == "coletasaprovar") {
                 const result = await db.query("DELETE FROM coletasaprovar WHERE id = ($1) RETURNING *", [req.body.itemId]);
-                await db.query(`INSERT INTO coletasdeletadas (company, date, product, username, volume, weight, "order") VALUES($1, $2, $3, $4, $5, $6, $7)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order]);
+                await db.query(`INSERT INTO coletasdeletadas (company, date, product, username, volume, weight, "order", loja) VALUES($1, $2, $3, $4, $5, $6, $7, $8)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order, result.rows[0].loja]);
                 res.status(200).send("ok");
             }
         }
@@ -179,15 +186,20 @@ async function collects(app, _) {
             if (companiesArray.includes(req.body.input.company)) {
                 if (dateRegex.test(req.body.input.date)) {
                     if (productsArray.includes(req.body.input.product)) {
-                        if (req.body.filter == "coletas") {
-                            await db.query(`UPDATE coletas SET company = $1, date = $2, product = $3, username = $4, volume = $5, weight = $6, "order" = $7 WHERE id = $8`, [req.body.input.company, req.body.input.date, req.body.input.product, req.body.collaborator.name, req.body.input.volume, req.body.input.weight, req.body.input.order, req.body.input.id ]);
+                        if (firmArray.includes(req.body.input.loja)) {
+                            if (req.body.filter == "coletas") {
+                            await db.query(`UPDATE coletas SET company = $1, date = $2, product = $3, username = $4, volume = $5, weight = $6, "order" = $7, loja = $8 WHERE id = $9`, [req.body.input.company, req.body.input.date, req.body.input.product, req.body.collaborator.name, req.body.input.volume, req.body.input.weight, req.body.input.order, req.body.input.loja, req.body.input.id]);
                             const result = await db.query("DELETE FROM coletas WHERE id = ($1) RETURNING *", [req.body.itemId]);
-                            await db.query(`INSERT INTO coletasaprovar (company, date, product, username, volume, weight, "order") VALUES($1, $2, $3, $4, $5, $6, $7)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order]);
+                            await db.query(`INSERT INTO coletasaprovar (company, date, product, username, volume, weight, "order", loja) VALUES($1, $2, $3, $4, $5, $6, $7, $8)`, [result.rows[0].company, result.rows[0].date, result.rows[0].product, result.rows[0].username, result.rows[0].volume, result.rows[0].weight, result.rows[0].order, result.rows[0].loja]);
                             res.status(200).send("ok");
+                            }
+                            if (req.body.filter == "coletasaprovar") {
+                                await db.query(`UPDATE coletasaprovar SET company = $1, date = $2, product = $3, username = $4, volume = $5, weight = $6, "order" = $7, loja = $8 WHERE id = $9`, [req.body.input.company, req.body.input.date, req.body.input.product, req.body.collaborator.name, req.body.input.volume, req.body.input.weight, req.body.input.order, req.body.input.loja, req.body.input.id]);
+                                res.status(201).send("ok");
+                            }
                         }
-                        if (req.body.filter == "coletasaprovar") {
-                            await db.query(`UPDATE coletasaprovar SET company = $1, date = $2, product = $3, username = $4, volume = $5, weight = $6, "order" = $7 WHERE id = $8`, [req.body.input.company, req.body.input.date, req.body.input.product, req.body.collaborator.name, req.body.input.volume, req.body.input.weight, req.body.input.order, req.body.input.id]);
-                            res.status(201).send("ok");
+                        if (!firmArray.includes(req.body.input.loja)) {
+                            res.status(404).send("Coloque um nome de loja válido");
                         }
                     }
                     if (!productsArray.includes(req.body.product)) {
